improved the binding API and added comments
